implements react router and adds photoBooth component
