+++ src/App.js
@@ -1,80 +1,12 @@
 import React, { Component } from 'react'
-import PhotoList from './PhotoList'
-
-const COUNTDOWN = 5
+import { Link } from 'react-router-dom';
 
 class App extends Component {
   constructor(props) {
     super(props)
     this.state = {
-      width: 300,
-      height: 0,
-      streaming: false,
-      photos: [],
-      counter: COUNTDOWN,
       photoQuantity: 0,
       errorMessage: '',
-    }
-  }
-
-  componentDidMount() {
-    const video = document.getElementById('video');
-    const canvas = document.getElementById('canvas');
-
-    video.addEventListener('canplay', () => {
-      if (!this.state.streaming) {
-        const height = video.videoHeight / (video.videoWidth/this.state.width)
-
-        this.setState({ height })
-
-        video.setAttribute('width', this.state.width);
-        video.setAttribute('height', this.state.height);
-        canvas.setAttribute('width', this.state.width);
-        canvas.setAttribute('height', this.state.height);
-        this.setState({streaming: true});
-      }
-    })
-
-    this.countDown()
-  }
-
-  countDown = () => {
-    setInterval(() => {
-      if (this.state.streaming && this.state.counter > 0) {
-        this.setState({ counter: this.state.counter - 1 })
-      } else if (this.state.counter === 0) {
-        if (this.state.photos.length < this.state.photoQuantity) this.takePicture()
-        this.setState({ counter: COUNTDOWN })
-      }
-    }, 1000)
-  }
-
-  startStream = () => {
-    const video = document.getElementById('video');
-
-    navigator.mediaDevices.getUserMedia({ video: true, audio: false })
-      .then((stream) => {
-        video.srcObject = stream;
-        video.play();
-      })
-      .catch((err) => {
-        console.log('An error occured! ' + err);
-      });
-  }
-
-  takePicture = () => {
-    const canvas = document.getElementById('canvas');
-    const context = canvas.getContext('2d');
-    const video = document.getElementById('video');
-
-    if (this.state.width && this.state.height) {
-      canvas.width = this.state.width;
-      canvas.height = this.state.height;
-      context.drawImage(video, 0, 0, this.state.width, this.state.height);
-
-      const data = canvas.toDataURL('image/png');
-
-      this.setState({ photos: [...this.state.photos, data] })
     }
   }
 
@@ -98,27 +30,9 @@ class App extends Component {
   render() {
     return (
       <div className='app'>
-        <div className='video-container'>
-            <p
-            className='counter font-mono text-md text-grey-darkest'
-              hidden={!this.state.streaming || this.state.photos.length === this.state.photoQuantity}
-            >
-              {this.state.counter}
-            </p>
-            <video id='video' className={this.state.streaming ? 'video-capturing' : 'video-hidden'}>
-              Video stream not available.
-            </video>
-            <p
-              className='counter font-mono text-grey-darkest'
-              hidden={!this.state.streaming || this.state.photos.length === this.state.photoQuantity}
-            >
-              {this.state.counter}
-            </p>
-          </div>
           <div className='enter-booth-container'>
             <h1
               className='font-mono text-md text-grey-darkest text-center'
-              hidden={this.state.streaming}
             >
               WebRTC Photo Booth
             </h1>
@@ -126,27 +40,22 @@ class App extends Component {
               className='number-of-photos font-mono text-grey-darkest text-center'
               placeholder='Enter Number Of Photos'
               onChange={(e) => this.validatePhotoQuantity(e)}
-              hidden={this.state.streaming}
             />
             <p
               className='error-message font-mono text-md text-red'
             >
               {this.state.errorMessage}
             </p>
+            <Link to='/photoBooth'>
             <button
               id='startbutton'
-              hidden={this.state.streaming}
-              onClick={this.startStream}
               className='font-mono text-md bg-teal hover:bg-teal-dark text-white mx-auto p-4 rounded'
               disabled={!this.state.photoQuantity > 0 && this.state.photoQuantity <= 5}
             >
                 Enter Photo Booth
             </button>
+            </Link>
           </div>
-        <canvas hidden={true} id='canvas'/>
-        <div className='output'>
-          <PhotoList photos={this.state.photos} countDown={this.countDown}/>
-        </div>
       </div>
     );
   }
